--- a/btf/btf_test.go
+++ b/btf/btf_test.go
@@ -313,6 +313,12 @@ func TestLoadSpecFromElf(t *testing.T) {
 			t.Errorf("Expected Void for type id 0, but got: %T", vt)
 		}
 
+		if id, err := spec.TypeID(vt); err != nil {
+			t.Error("Error when looking up type id for Void:", err)
+		} else if id != 0 {
+			t.Error("Expected Void to have id 0, got", id)
+		}
+
 		var bpfMapDef *Struct
 		if err := spec.TypeByName("bpf_map_def", &bpfMapDef); err != nil {
 			t.Error("Can't find bpf_map_def:", err)