@lmb 2023-03-28
Closing this for now. The gain really isn't that big.